refactor(var): rename utility files and improve code organization
