+++ packages/utilities/var/src/construction-detection.ts
@@ -3,7 +3,7 @@ import type { Scope } from "@typescript-eslint/scope-manager";
 import type { TSESTree } from "@typescript-eslint/types";
 
 import { AST_NODE_TYPES as T } from "@typescript-eslint/types";
-import { getVariableDefinitionNode } from "./variable-resolver";
+import { getVariableDefinitionNode } from "./get-variable-definition-node";
 
 export const ConstructionDetectionHint = {
   None: 0n,
