+++ packages/utilities/var/src/get-variables-from-scope.ts
