+++ packages/utilities/var/src/index.ts
@@ -1,7 +1,7 @@
-export * from "./misc";
-export * from "./value-construction";
-export * from "./value-equal";
-export * from "./variable-assignment";
-export type * from "./variable-declaration";
-export * from "./variable-extractor";
-export * from "./variable-resolver";
+export * from "./construction-detection";
+export * from "./find-assignment-target";
+export * from "./find-property";
+export * from "./get-child-scopes";
+export * from "./get-variable-definition-node";
+export * from "./get-variables-from-scope";
+export * from "./is-node-value-equal";
